image-processor/requirements.txt
Flask==1.1.1
opencv-python-headless==4.4.0.46
numpy==1.18.5
tensorflow-gpu==2.3.0
easydict==1.9
Pillow==8.1.0
cloudevents==1.2.0
requests==2.25.1